refactor: move homepage selection to /helpers as it should not be at the root of /home and it is a helper hook
