+++ src/features/home/helpers/selectHomepageEntry.test.ts
@@ -3,7 +3,7 @@ import shuffle from 'lodash/shuffle'
 import { EligibilityType, UserProfileResponse, UserRole } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { useUserHasBookings } from 'features/bookings/api/useUserHasBookings'
-import { useSelectHomepageEntry } from 'features/home/selectHomepageEntry'
+import { useSelectHomepageEntry } from 'features/home/helpers/selectHomepageEntry'
 import { Credit, getAvailableCredit } from 'features/user/helpers/useAvailableCredit'
 import { HomepageEntry, Tag } from 'libs/contentful'
 import { useRemoteConfigContext } from 'libs/firebase/remoteConfig'
